requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
faiss-cpu
omegaconf>=2.3.0
numpy
matplotlib
Pillow